+++ src/board.h
@@ -48,7 +48,7 @@
 uint64_t board_flip(uint64_t base, uint8_t x, uint8_t y);
 uint64_t board_set_high(uint64_t base, uint8_t x, uint8_t y);
 uint64_t board_set_low(uint64_t base, uint8_t x, uint8_t y);
-int board_get(uint64_t base, uint8_t x, uint8_t y);
+uint8_t board_get(uint64_t base, uint8_t x, uint8_t y);
 void board_setup(uint64_t pos_boards[12], uint64_t occupancy_boards[3]);
 
 #endif
